backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic>=2.10.3
pydantic-settings>=2.4.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
google-generativeai==0.3.2
slowapi==0.1.8
langchain==0.3.23
langchain-community==0.3.21
unstructured-client==0.32.1